requirements.txt
streamlit==1.31.1
anthropic==0.3.11
httpx==0.24.1
python-dotenv==1.0.0